+++ lib/app_caption.dart
@@ -1,11 +1,15 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:money/helpers/date_helper.dart';
 import 'package:money/models/settings.dart';
 import 'package:money/views/view_pending_changes/bage_pending_changes.dart';
+import 'package:money/widgets/gaps.dart';
 
 class AppCaption extends StatelessWidget {
-  final String subCaption;
+  final Widget child;
 
-  const AppCaption({super.key, required this.subCaption});
+  const AppCaption({super.key, required this.child});
 
   @override
   Widget build(BuildContext context) {
@@ -21,13 +25,48 @@ class AppCaption extends StatelessWidget {
             itemsDeleted: Settings().trackMutations.deleted,
           )
         ]),
-        Text(
-          subCaption,
-          textAlign: TextAlign.left,
-          overflow: TextOverflow.ellipsis,
-          style: const TextStyle(fontSize: 10),
-        ),
+        child,
       ],
+    );
+  }
+}
+
+class LoadedDataFileAndTime extends StatelessWidget {
+  final String filePath;
+  final DateTime? lastModifiedDateTime;
+
+  const LoadedDataFileAndTime({
+    super.key,
+    required this.filePath,
+    required this.lastModifiedDateTime,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      reverse: true,
+      // controller: _scrollController,
+      scrollDirection: Axis.horizontal,
+      dragStartBehavior: DragStartBehavior.down,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: [
+          Text(
+            filePath,
+            textAlign: TextAlign.left,
+            overflow: TextOverflow.ellipsis,
+            style: const TextStyle(fontSize: 10),
+          ),
+          gapMedium(),
+          Text(
+            geDateAndTimeAsText(lastModifiedDateTime),
+            textAlign: TextAlign.left,
+            overflow: TextOverflow.ellipsis,
+            style: const TextStyle(fontSize: 12, fontWeight: FontWeight.bold),
+          ),
+          gapMedium(),
+        ],
+      ),
     );
   }
 }