+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [version 1.7.4] 2024-03-05
 
 ### Added
-
+- Show last modified date of opened file
 - Show currency column in Transaction view
 - Custom Filter of column by currency